// src/plugins/agent_traces/public/application/utils/interfaces.ts
export interface ChartDataBucket {
  key: number | string;
  doc_count: number;
}

/**
 * Interface for Ordered
 */
interface Ordered {
  date: true;
  interval: Duration;
  intervalOpenSearchUnit: string;
  intervalOpenSearchValue: number;
  min: Moment;
  max: Moment;
}

/**
 * Interface for histogram series data
 */
export interface HistogramSeries {
  id: string;
  name: string;
  data: Array<{ x: number; y: number }>;
}

/**
 * Interface for chart data
 */
export interface ChartData {
  values: Array<{ x: number | string; y: number }>;
  xAxisOrderedValues: Array<number | string>;
  xAxisFormat: { id: string; params: { pattern: string } };
  xAxisLabel: string;
  yAxisLabel: string;
  buckets?: ChartDataBucket[];
  ordered: Ordered;
  series?: HistogramSeries[];
}

/**
 * Interface for bucket interval
 */
export interface BucketInterval {
  scaled?: boolean;
  description?: string;
  scale?: number;
  interval?: string;
}
// ...
export interface BaseProcessedSearchResults {
  hits: ISearchResult['hits'];
  fieldCounts: Record<string, number>;
  dataset: Dataset;
  elapsedMs: number;
}

/**
 * Interface for processed search results
 */
export interface ProcessedSearchResults extends BaseProcessedSearchResults {
  chartData?: ChartData;
  bucketInterval?: BucketInterval;
}

/**
 * Interface for Traces Chart processor processed search results
 */
export interface TracesChartProcessedResults extends BaseProcessedSearchResults {
  requestChartData?: ChartData;
  errorChartData?: ChartData;
  latencyChartData?: ChartData;
  bucketInterval?: BucketInterval;
}